+++ activerecord/test/models/treasure.rb
@@ -1,6 +1,6 @@
 class Treasure < ActiveRecord::Base
   has_and_belongs_to_many :parrots
-  belongs_to :looter, :polymorphic => true
+  belongs_to :looter, :polymorphic => true, :poly_classes => [Pirate, Parrot]
   belongs_to :ship
 
   has_many :price_estimates, :as => :estimate_of